requirements.txt
allure-python-commons~=2.14.2
selenium~=4.33.0
pytest~=8.4.0
Faker~=37.3.0